pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "kernel-foundry"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.50",
    "python-dotenv>=1.0",
    "torch>=2.1",
    "triton>=3.0",
    "numpy>=1.26",
    "pyyaml>=6.0",
]

[project.scripts]
kernel-foundry = "main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["kernel_foundry*", "tasks*"]
